el8toel9: Warn about deprecated Xorg drivers
Some Xorg drivers have been deprecated in favor of the "modesetting"
driver.

If Xorg is configured to use those driver, it may not be able to work
properly after the upgrade.

Add a new actor to check in the journal whether such Xorg drivers are in
use and in that case, also warn if there are custom Xorg config options.

Known limitation: This actor checks the journal logs since the last
boot, meaning that if Xorg was started manually from a console or if the
system has been rebooted after a graphical Xorg session was used, the
actor will not be able to detect the use of deprecated drivers.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>