Fix the error msg for DNF missing diskspace
Currently we inform users they do not have enough space on
a partition hosting the /var/lib/leapp directory when our dnf upgrade
plugin ends with non-zero exit code and the err msg contains string
like:
    At least <size> more space needed on the <path> filesystem.

This is true only during phases prior the reboot when a partition is
not represented by disk images, which we create to handle partitions
with XFS filesystems without ftype attributes. Such images have size
2GiB unless the LEAPP_OVL_SIZE envar is specified.

In general, the real missing space is currently not known during
the phases prior the reboot as we "emulate" the original storage using
overlayfs. So let's provide better correct error msg for these
disk space errors during phases prior the reboot.

But during the upgrade phase, when the DNF transaction is going to
be performed, the host storage is not exposed to the DNF plugin and
so error messages informing about needed more space are really
correct - in such a case, does not try to interpret the error msg
in any additional way and expose raw data about the error.

tl;dr;
* Switch to the generic error msg always during the upgrade phase,
  exposing the raw error msg without trying to interpret it.
* Interpret correctly the error msg for such disk space issues
  during phases before the system reboot, when the real host storage
  is emulated.