dnfplugin: Switch to generic(old) err msg for diskspace info during the upgrade
Currently we inform users they do not have enough space on
a partition hosting the /var/lib/leapp directory when our dnf upgrade
plugin ends with non-zero exit code and the err msg contains string
like:
    At least <size> more space needed on the <path> filesystem.

This is partially true during phases prior the reboot as we "emulate"
the original storage using overlayfs. But during the upgrade phase,
when the DNF transaction is going to be performed, the message is
missleading as the host storage is not exposed to the plugin and so
the error messages informing about needed more space are really
correct.

* Switch to the generic error msg always during the upgrade phase,
  exposing the raw error msg without trying to interpret it.
* Interpret the error msg for such disk space issues during phases
  before the system reboot, when the real host storage is emulated.