update URLs to docs: fix links and use shortened URLs when needed
It happened several times that a link we used in reports became invalid
over time, so it could not provide expected help to customers. The
fix of such links require update of the code, which does not have to be
possible on existing releases. Instead of that, we rather decided
to use shortnened URLs when possible and when it makes sense to reduce
the risk. The benefit is that we are able to change the target
destination of such https://red.ht/... URLs so we can fix such issues
without the need to update the code (and rpms).

New shortened URLs:
  * https://red.ht/upgrading-rhel7-to-rhel8-main-official-doc
  * https://red.ht/upgrading-rhel8-to-rhel9-main-official-doc
  * https://red.ht/ipu-customisation-repos-known-issues
  * https://red.ht/leapp-dialogs
  * https://red.ht/file-systems-and-storage-removed-btrfs-rhel-8
  * https://red.ht/rhel-8-migrate-postgresql-server
  * https://red.ht/rhel-8-python
  * https://red.ht/rhel-8-configuring-routing-protocols
  * https://red.ht/nm-settings-ifcfg-rh
  * https://red.ht/rhel-9-intel-microarchitectures
  * https://red.ht/rhel-9-blog-openssh-scp-deprecation

Additional changes:
  * checktargetrepos: replace "TBA" by valid URL for 8 -> 9 docs
  * dropped unneeded pylint skips