Experimental (unsupported) enable of upgrades for encrypted devices
Upgrades with encrypted devices / partitions are not supported now.
Based on the feedback from @Naranthiran, it is possible to upgrade
with encrypted devices (using TPM) with the following changes.

We have no idea in what all situations this will work correctly,
nor detection of used technologies is implemented to say whether
the decryption of encrypted devices will happen automatically
correctly or not, etc. From this point, allowing the upgrade
in unsupported mode when following devel envar is set:
    LEAPP_DEVEL_ALLOW_DISK_ENCRYPTION=1

For more details, see:
  https://github.com/oamg/leapp-repository/discussions/1098

Authored / Co-authored-by @Naranthiran