overlay lib: Secure the creation of the disk image when size is < 130
In case the filesystem for which the disk img is going to be created
has very small amount of free space (under 130 MiBs), it cannot be
formatted by XFS with current params.

This could be hit in several cases:
 * the system partition/volume - in this case, most likely an issue
   will be hit anyway later by DNF speaking about small amount of
   free space if a content is installed inside by RPMs
   as such a small amount of free space is really not expected
   to see at all
 * it's a data mount point (e.g. iso) or a filesystem type that should
   by part of the OVERLAY_DO_NOT_MOUNT set, so enlarging the value
   to 130 MiBs should not affect anything negatively at all
 * in case of /boot, the problem with the free space is covered
   already in a different actor prior we try to create any disk img,
   so we are safe here

Based on arguments above, I am considering setting the 130 MiBs as
minimal value safe for in-place upgrades. Also it will allow to skip
possible problems with specific file systems (like tmpfs, ...) in case
we are still missing some in the OVERLAY_DO_NOT_MOUNT - and kind of
read only storage (such as iso9660, etc..).

Co-authored-by: Michal Hečko <michal.sk.com@gmail.com>