pcidevicescanner: skip test_actor_execution if lspci is not installed
Checking the test it does not have so big value. The original
functionality has been extended and the particular tests does not
reflect it. Majority of the functionality is tested in unit tests
around, so regarding that, I prefer for now to skip the test if
lspci utility is not installed and get back to it later.