updategrubcore: Skip on IBM Z to omit confusing report
As GRUB2 is not present on IBM Z, related commands cannot be used
on this architecture as they just fail (as the executors are not
present). In this particular case, logs contain a confusing traceback
msg, which is not harmful, but it's not nice niether. Added check
for s390x architecture.

Note that similar problem could possibly appear on baremetal POWER machines,
where petitboot is used. After the fix of the leapp run functions
from the stdlib, catch also OSError exception. In such a case, print
info the grub2-core is not present with the hint that's most likely
it should not be even called on the particular machine and and the
actor execution too.

Suppress deprecated warnings in unit tests for get_grub_device() tests.