load data files: update hints in error messages and prepare for the download drop
As the leapp upgrade data files are nowadays part of the install rpm,
there is no need to download them anymore. Also, we plan to drop the
service providing the data files online in future.

For that reason, update all texts and related error messages so people
are not instructed to visit obsoleted article and do not try to apply
invalid (obsoleted) data files anymore.

Right now, we are keeping the functionality for the data files download,
but the fetch functino is already updated and prepared to stop trying
to download the files if not present. As we have the functionality
already present, I think we should keep a possibility of the download
for additional custom data files (not provided by us) in custom actors,
but for the official data files we will require them in future to be
present locally only.

NOTE: regarding another planned changes soon, skipping update of
unit-tests

Co-authored-by: Michal Hečko <michal.sk.com@gmail.com>