load data files: do not try to download data files when missing
As the leapp upgrade data files are nowadays part of the install rpm,
there is no need to download them anymore. Also, we do not have any
plans to provide the updated data files anywhere out of the rpm.

From that point, do not try to download provided data files anymore.