Introduce custom modifications tracking
This commit introduces two actors:
* the scanner that that scans leapp files and
  produces messages with actor name/filepath mapping in case
  any unexpected custom files or modified files were discovered.
* the checker that processes CustomModification messages and
  produces report entries.

* uses rpms.get_leapp_packages function
* pstodulk: Updated report messages to provide more information to users

The purpose of this change is to help with the investigation
of reported issues as people harm themselves from time to time
and as this is not usually expected, it prolongs the solution
of the problem (people investigating such issues do not check
this possibility as the first thing, which is understandable).
This should help to identify possible root causes faster as
report msg should be always visible.

Jira: RHEL-1774

Co-authored-by: Petr Stodulka <pstodulk@redhat.com>