upgrade data files loading: update error msgs and repors + minor changes
Updated number of error messages and reports to be sure that users
know what files are actually problematic. Original errors and reports
usually didn't contain the full path to an upgrade data file due to
possibility to download the file from a server. However, the posibility
to download fresh data files from a requested server is not expected
to support in the current state as the data files are part of provided
packages.

I've been thinking quite long time whether to actually drop
or deprecate bigger part of the code to simplify the whole solution,
as currently it's not planned to have a possibility to download
some data files from servers in future. However, thinking about
upcoming challenges, I am not totally persuaded that we will not
revive that functionality in future, or that we will not want to
use it for something little bit different. From that POV (and late
phase of development prior the planned release) I think it will be
better to preserve it for now and raise a discussion about it later.

Other changes in this PR:
 * drop hardcoded name of the leapp-upgrade-elXtoelY rpm and use
   the leapp.libraries.common.rpms.get_leapp_packages() function
 * replace REPOSITORY group by SANITY; it was originally a mixture
   of both and SANITY fits better to me from this point
 * the check of consumed data sets could produce report with empty
   links, as the original article(s) we referred to have been obsoleted;
   so added filter for missing URLs

Co-authored-by: Toshio Kuratomi <a.badger@gmail.com>