device driver deprecation data: print nice error msg on malformed data
In case of malformed device_driver_deprecation_data.json user could
originally see raw traceback without having too much information
what it actually means or how to fix it. That usually happens only
when the file is manually modified on the machine. So in this case
we inform user what file is problematic and how to restore the original
file installed by our package.

In case of upstream development, this msg could be seen also when
new data is provided if:
 * data file is malformed
 * data file has a new format of data (still json expected)
 * etc.
These issues however will be discovered prior the merge as the
running tests will fail, so such a problematic file should never
get part of the upstream. From that point, we will be expect that
user has malformed / customized data file. So no need to handle
all possible errors differently in this case.

Jira: OAMG-7549

Co-authored-by: Toshio Kuratomi <a.badger@gmail.com>