transitionsystemdservicesstates: Do not handle libvirt and virt* services on 8->9+
Beginning with RHEL8 there are two possible configurations (monolithic
and modular) for the libvirt and virt* (virtqemud, virtlockd,
virtproxyd, ...) services. The generic strategy implemented in the
transitionsystemdservicesstates actor cannot handle such setup
and leaves the upgraded system with a an incorrect combination of these
configurations. Therefore these services are ignored by the actor and
their state transition will be done by the services (and other related
units) themselves.

Jira: RHEL-22597