Add artificial 9->10 repomaping data
Note that the data is artificial, meaning it's not generated through
repomapping generator and therefore not verified by it against the
source of truth. The assumption is that the repos will be similar to how
they are on RHEL 9 and are basically copied (redundant channels and rhui
repos are dropped).