IPU 9 -> 10: Migrate & rebuild RPM DB
The RPM DB has been moved from /var/lib/rpm to /usr/lib/sysimage/rpm
in RHEL 10. Apply the change and create symlink in the original
path to the new one as expected.

Also rebuild the RPM DB to ensure it's compatible with the new RPM
version. Previously the RPM DB was being rebuilt during IPU 8 -> 9.
However after discussion with RPM SMEs it has been decided that this
is actually very reasonable to do always. So applying for any upgrade
path but IPU 7 -> 8 (let's do not change this one so much when it's
kind of finished).

Co-authored-by: Matej Matuska <mmatuska@redhat.com>