Update reboot msg: Note the console access
Some users hasn't read the upgrade documentation and are not aware
that after the reboot the actual upgrade is processing. As they wait
just for the ssh connection, they think that something is wrong and
sometimes reboot the machine, interrupting the entire process, making
the machine broken in some cases.

Adding info that they need a console access in case they want to watch
the upgrade progress.

Jira: https://issues.redhat.com/browse/RHEL-27231