ipa actor: add shortened url for 9-to-10 migration guide
During the RHEL 9.5 to RHEL 10.0 Leapp preupgrade process, the
check_ipa_server actor crashes due to a KeyError while trying to
find the URL for the 9-to-10 migration guide.

Add a URL for the key '9' and make the code more robust by
defaulting to "TBD" if the key is not found in the dictionary
storing version/url-of_migration guide.

Fixes: https://issues.redhat.com/browse/RHEL-50829
Co-authored-by: Vojtech Sokol <vsokol@redhat.com>
Signed-off-by: Florence Blanc-Renaud <flo@redhat.com>